requirements.txt
asgiref==3.5.2
Django==4.1.4
djangorestframework==3.14.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
